+++ apps/admin-x-settings/src/components/settings/site/designAndBranding/GlobalSettings.tsx
@@ -9,15 +9,18 @@ import {getImageUrl, useUploadImage} from '@tryghost/admin-x-framework/api/image
 import {useFramework} from '@tryghost/admin-x-framework';
 import {useGlobalData} from '../../../providers/GlobalDataProvider';
 import {useHandleError} from '@tryghost/admin-x-framework/hooks';
-import type {Font} from '@tryghost/custom-fonts';
+import type {Font, HeadingFont} from '@tryghost/custom-fonts';
 
 // TODO: create custom types for heading and body fonts in @tryghost/custom-fonts, so we can extend
 // them separately
 type BodyFontOption = {
-    value: Font | 'Theme default',
-    label: Font | 'Theme default'
+    value: Font | typeof DEFAULT_FONT,
+    label: Font | typeof DEFAULT_FONT
 };
-type HeadingFontOption = BodyFontOption;
+type HeadingFontOption = {
+    value: HeadingFont | typeof DEFAULT_FONT,
+    label: HeadingFont | typeof DEFAULT_FONT
+};
 
 export interface GlobalSettingValues {
     description: string
@@ -46,10 +49,10 @@ const GlobalSettings: React.FC<{ values: GlobalSettingValues, updateSetting: (ke
 
     // TODO: replace with getCustomFonts() once custom-fonts is updated and differentiates
     // between heading and body fonts
-    const customHeadingFonts: HeadingFontOption[] = CUSTOM_FONTS.map(x => ({label: x, value: x}));
+    const customHeadingFonts: HeadingFontOption[] = CUSTOM_FONTS.heading.map(x => ({label: x, value: x}));
     customHeadingFonts.unshift({label: DEFAULT_FONT, value: DEFAULT_FONT});
 
-    const customBodyFonts: BodyFontOption[] = CUSTOM_FONTS.map(x => ({label: x, value: x}));
+    const customBodyFonts: BodyFontOption[] = CUSTOM_FONTS.body.map(x => ({label: x, value: x}));
     customBodyFonts.unshift({label: DEFAULT_FONT, value: DEFAULT_FONT});
 
     const selectedHeadingFont = {label: headingFont, value: headingFont};
